requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
asyncpg==0.30.0
iso8601==2.1.0
mysql-connector-python==9.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.2.1
python-dotenv==1.0.1
pytz==2024.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
